move validate to own file
